@@ -2,29 +2,28 @@
   "name": "clipp.li",
   "author": "Jonathan Wieben",
   "license": "MIT",
-  "engines": {
-    "node": "10.x"
+  "scripts": {
+    "dev": "next",
+    "build": "next build",
+    "start": "next start"
   },
   "dependencies": {
     "axios": "^0.19.0",
-    "body-parser": "^1.19.0",
     "classnames": "^2.2.6",
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "grommet": "^2.7.8",
+    "grommet": "^2.8.1",
     "moment": "^2.24.0",
-    "mongodb": "^3.3.2",
-    "next": "^9.0.6",
+    "mongodb": "^3.3.5",
+    "next": "^9.1.4",
     "prop-types": "^15.7.2",
     "randomstring": "^1.1.5",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "styled-components": "^4.4.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "styled-components": "^4.4.1",
     "url": "^0.11.0"
   },
   "devDependencies": {
-    "eslint": "^6.4.0",
-    "eslint-plugin-react": "^7.14.3",
-    "prettier": "^1.18.2"
+    "eslint": "^6.7.2",
+    "eslint-plugin-react": "^7.17.0",
+    "prettier": "^1.19.1"
   }
 }
